+++ requirements.txt
@@ -3,6 +3,7 @@ blinker==1.8.2
 click==8.1.7
 colorama==0.4.6
 Flask==3.0.3
+Flask-JWT-Extended==4.6.0
 greenlet==3.1.1
 gunicorn==23.0.0
 iniconfig==2.0.0
@@ -15,6 +16,7 @@ pandas==2.2.3
 passlib==1.7.4
 pluggy==1.5.0
 psycopg2==2.9.9
+PyJWT==2.9.0
 pypdf==5.0.0
 pytest==8.3.3
 python-dateutil==2.9.0.post0
